Add frame pointer unwinding for aarch64
If we don't find any debug information for a given frame, we usually
cannot unwind any further. However, the binary in question might have
been compiled with frame pointers, in which case we can look up the
well known frame pointer locations in the stack snapshot and use them
to bridge the frames without debug information.

The unwinding on aarch64 is pretty similar to the one on arm, the
difference is in the offsets of LR and FP on the stack.

Signed-off-by: Ulf Hermann <ulf.hermann@qt.io>